added suffixes to geocodes dictionaries
added letter suffixes to geocodes to allow for multiple geographies at a geocode level. It's just a quick fix to be able to add 'county subdivision' to the dictionary as a level 3 which is the same as 'tract'. Need to explore ways in the future to accept an alternative geocode hierarchy dictionary to expand the types of geographies accepted by the "ConstructGeoClauses" function.